% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanova_fit.R
\name{fanova_components}
\alias{fanova_components}
\title{(Internal) Extract component functions}
\usage{
fanova_components(object)
}
\arguments{
\item{object}{A fanova object.}
}
\value{
A matrix where each column is an evaluated component function.
}
\description{
(Internal) Extract component functions
}
