% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanova.R
\name{functionalANOVA}
\alias{functionalANOVA}
\title{generalized functional analysis of variance}
\usage{
functionalANOVA(data, vars, n = c(10, 2), model,
  predict.fun = function(object, newdata) predict(object, newdata = newdata),
  weight.fun = NULL)
}
\arguments{
\item{data}{a \code{data.frame} or \code{data.table} which contains the features/covariates on which \code{predict.fun} was learned/estimated.}

\item{vars}{a \code{character} vector which corresponds to a subset of the columns of \code{data}.}

\item{n}{a \code{numeric} vector of length 2, where the first element corresponds to the dimension of the grid constructed for each of the elements of \code{vars} and the second element to the number of rows to sample from \code{data}.}

\item{model}{the first argument to \code{predict.fun}, presumably a model object which can make predictions.}

\item{predict.fun}{a \code{function} whose first two arguments are "object" and "newdata" which returns a numeric vector the same length as the number of rows in newdata. the default value is to call the \code{predict} method on the \code{model}.}

\item{weight.fun}{a \code{function} with two arguments, \code{design} and \code{data}, both of which are \code{data.tables} which returns a \code{numeric} of the same length as the number of rows in \code{design}. this is intended for use to use the \code{data} to estimate the distribution of the input features, and then to use that estimate to the probability of points in the \code{design} grid.}
}
\value{
a \code{data.table} with columns for a grid of points of the \code{vars}, a (set of) column(s) that correspond to the estimated effect of those features/covariates on the prediction function, and a column \code{effect} which indicates which subset of the covariates/features each estimate belongs to.
}
\description{
computes a generalized (weighted) functional ANOVA decomposition of a prediction function, giving the best additive decomposition of the prediction function in terms of squared error.
}
\references{
Giles Hooker. Generalized Functional ANOVA Diagnostics for High Dimensional Functions of Dependent Variables, Journal of Computational and Graphical Statistics, Vol. 16, No. 3 (2007), pp. 709-732.
}
