% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fanova.R
\name{plot.fanova}
\alias{plot.fanova}
\title{Plot fanova importance}
\usage{
\method{plot}{fanova}(x, ...)
}
\arguments{
\item{x}{A 'fanova' object.}

\item{...}{Other arguments passed to \code{barplot}.}
}
\value{
Invisibly returns the result of \code{barplot}.
}
\description{
Creates a bar plot of the variance shares for each component in the
fANOVA decomposition.
}
