% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.fanova.R
\name{summary.fanova}
\alias{summary.fanova}
\alias{print.summary.fanova}
\title{Summarize a fanova fit}
\usage{
\method{summary}{fanova}(object, ...)

\method{print}{summary.fanova}(x, ...)
}
\arguments{
\item{object}{A 'fanova' object.}

\item{...}{Other arguments (unused).}
}
\value{
An object of class 'summary.fanova', which is a list containing
  the model call, dimensions, and the importance data frame.
}
\description{
Provides a summary of the fitted functional ANOVA model, including the
variance decomposition.
}
